+++ pioreactor/tests/conftest.py
@@ -51,7 +51,7 @@ def run_around_tests(request):
         c.pop(test_name)
     with local_persistent_storage("alt_media_fraction") as c:
         c.pop(test_name)
-    with local_persistent_storage("liquid_volume") as c:
+    with local_persistent_storage("current_volume_ml") as c:
         c.pop(test_name)