+++ update_scripts/upcoming/update.sh
@@ -22,6 +22,11 @@ if [ "$HOSTNAME" = "$LEADER_HOSTNAME" ]; then
     crudini --set /home/pioreactor/.pioreactor/config.ini stirring.config initial_target_rpm $(crudini --get /home/pioreactor/.pioreactor/config.ini stirring.config target_rpm)
     crudini --del /home/pioreactor/.pioreactor/config.ini stirring.config target_rpm
 
+    # update max_volume_ml to max_working_volume_ml
+    crudini --set /home/pioreactor/.pioreactor/config.ini bioreactor max_working_volume_ml $(crudini --get /home/pioreactor/.pioreactor/config.ini bioreactor max_volume_ml)
+    crudini --del /home/pioreactor/.pioreactor/config.ini bioreactor max_volume_ml
+
+    # TODO: confirm I don't need to update 04_dosing_automation.yaml, and the 08_liquid_volume.yaml chart
 
     # add raw od readings to export
     EXPORTABLE_DATASETS="/home/pioreactor/.pioreactor/exportable_datasets"